requirements.txt
autopep8==2.0.1
certifi==2022.12.7
flake8==6.0.0
gitdb==4.0.10
GitPython==3.1.31
mccabe==0.7.0
mypy==1.0.1
mypy-extensions==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-box==7.0.1
smmap==5.0.0
typing_extensions==4.5.0
python-box[all]~=7.0.1